Refactor getFees function and connect it to Tools util
